requirements.txt
kaggle>=1.6.0
pyarrow>=14.0.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
great-expectations>=1.0.0
mlflow>=2.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
pytest>=7.4.0
